Update Source Code

+++ JAPP/jp_promode.c
@@ -1,4 +1,3 @@
-
 #include "qcommon/q_shared.h"
 #include "bg_public.h"
 #include "bg_local.h"
@@ -7,7 +6,7 @@
 // Physics
 float	cpm_pm_airstopaccelerate = 1;
 float	cpm_pm_aircontrol = 0;
-float	cpm_pm_strafeaccelerate = 1.0;
+float	cpm_pm_strafeaccelerate = 1.0f;
 float	cpm_pm_wishspeed = 400;
 
 void CPM_UpdateSettings( int num ) {
@@ -17,14 +16,14 @@ void CPM_UpdateSettings( int num ) {
 	// Physics
 	cpm_pm_airstopaccelerate = 1;
 	cpm_pm_aircontrol = 0;
-	cpm_pm_strafeaccelerate = 1.0;
+	cpm_pm_strafeaccelerate = 1.0f;
 	cpm_pm_wishspeed = 400;
 	pm_accelerate = 10.0f;
-	pm_friction = 6.0;
+	pm_friction = 6.0f;
 
 	if ( num ) {
 		// Physics
-		cpm_pm_airstopaccelerate = 2.5;
+		cpm_pm_airstopaccelerate = 2.5f;
 		cpm_pm_aircontrol = 150;
 		cpm_pm_strafeaccelerate = 70;
 		cpm_pm_wishspeed = 30;
@@ -34,26 +33,28 @@ void CPM_UpdateSettings( int num ) {
 }
 
 void CPM_PM_Aircontrol( pmove_t *pmove, vector3 *wishdir, float wishspeed ) {
-	float	zspeed, speed, dot, k;
+	float zspeed, speed, dot, k;
 
-	if ( (pmove->ps->movementDir && pmove->ps->movementDir !=4) || wishspeed == 0.0) 
-		return; // can't control movement if not moveing forward or backward
+	// can't control movement if not moving forward or backward
+	if ( (pmove->ps->movementDir && pmove->ps->movementDir != 4) || (int)wishspeed == 0 )
+		return;
 
 	zspeed = pmove->ps->velocity.z;
 	pmove->ps->velocity.z = 0;
 	speed = VectorNormalize( &pmove->ps->velocity );
 
-	dot = DotProduct( &pmove->ps->velocity, wishdir);
-	k = 32; 
+	dot = DotProduct( &pmove->ps->velocity, wishdir );
+	k = 32;
 	k *= cpm_pm_aircontrol*dot*dot*pml.frametime;
-	
-	
-	if (dot > 0) {	// we can't change direction while slowing down
+
+
+	if ( dot > 0.0f ) {
+		// we can't change direction while slowing down
 		pmove->ps->velocity.x = pmove->ps->velocity.x*speed + wishdir->x*k;
 		pmove->ps->velocity.y = pmove->ps->velocity.y*speed + wishdir->y*k;
 		VectorNormalize( &pmove->ps->velocity );
 	}
-	
+
 	pmove->ps->velocity.x *= speed;
 	pmove->ps->velocity.y *= speed;
 	pmove->ps->velocity.z = zspeed;
